+++ pyramid_caching/cache.py
@@ -1,8 +1,10 @@
+from collections import namedtuple
+import hashlib
 import logging
 
 from zope.interface import implementer, classImplements
 
-from pyramid_caching.events import CacheHit, CacheMiss
+from pyramid_caching.events import ViewCacheHit, ViewCacheMiss
 from pyramid_caching.interfaces import (
     ICacheClient,
     ICacheManager,
@@ -67,30 +69,93 @@ class Manager(object):
         self.registry = registry
 
     def get_or_cache(self, get_result, prefixes, dependencies):
-        try:
-            versioned_keys = self.versioner.get_multi_keys(dependencies)
-        except CacheDisabled:
-            return get_result()
+        versioned_keys = self.versioner.get_multi_keys(dependencies)
+        key = CacheKey(prefixes, versioned_keys)
 
-        key_prefix = ':'.join(prefixes)
-        cache_key = key_prefix + ':' + ':'.join(versioned_keys)
+        cache_content = self.cache_client.get(str(key))
 
-        cached_result = self.cache_client.get(cache_key)
-
-        if cached_result is None:
+        if cache_content is not None:
+            result = self.serializer.loads(cache_content)
+            return CacheResult.hit(key, result)
+        else:
             result = get_result()
             data = self.serializer.dumps(result)
-            self.cache_client.add(cache_key, data)
-            self.registry.notify(CacheMiss(key_prefix))
-        else:
-            result = self.serializer.loads(cached_result)
-            self.registry.notify(CacheHit(key_prefix))
+            self.cache_client.add(str(key), data)
+            return CacheResult.miss(key, result)
 
-        return result
+
+class CacheKey(object):
+    """This object specifies the formatting of cache keys from a list of base
+    resource names (for example, a unique way to identify a Pyramid view
+    callable) and a list of names that define the current context of the view
+    (for example, a list of model names and the primary keys matching the
+    route).
+
+    ::
+       key = CacheKey(['mypackage.views', 'hello_view'], ['user:bob'])
+       key.root() --> 'mypackage.views:hello_view'
+       str(key) --> 'mypackage.views:hello_view:user:bob'
+
+    """
+    def __init__(self, bases, dependencies):
+        self.bases = bases
+        self.dependencies = dependencies
+
+    def root(self):
+        """The static part of the cache key that refers to a single view or
+        model class."""
+        return ':'.join(self.bases)
+
+    def key(self):
+        """The unique cache key identifying a resource and its context."""
+        return self.root() + ':' + ':'.join(self.dependencies)
+
+    def __str__(self):
+        return self.key()
+
+
+_CacheResultInfo = namedtuple("CacheResultInfo", ["key", "hit"])
+
+
+class CacheResult(object):
+    """This class represents a versioned cache response based on a unique
+    key."""
+    def __init__(self, cache_key, data, hit):
+        self._cache_key = cache_key
+        self.data = data
+        self._hit = hit
+
+    @classmethod
+    def hit(cls, cache_key, data):
+        """Specify that this data was fetched from the cache."""
+        return cls(cache_key, data, True)
+
+    @classmethod
+    def miss(cls, cache_key, data):
+        """Specify that this data was loaded from the application."""
+        return cls(cache_key, data, False)
+
+    def md5_hash(self):
+        """A string corresponding to the MD5 digest of the cache key that
+        uniquely defines a version of the data."""
+        m = hashlib.md5()
+        m.update(str(self._cache_key))
+        return m.hexdigest()
+
+    def info(self):
+        return _CacheResultInfo(self._cache_key, self._hit)
 
 
 class cache_factory(object):
+    """Decorator that defines the model dependencies for a view method.
 
+    ::
+       @cache_factory(depends_on={User: {'matchdict': 'user'}})
+       def hello_view(context, request):
+           user = User(request.matchdict['user'])
+           return "Hello, {}".format(user.name)
+
+    """
     def __init__(self, depends_on=None):
         self.depends_on = depends_on
 
@@ -106,8 +171,9 @@ class ViewCacheDecorator(object):
 
     def __call__(self, context, request):
         if not request.registry.settings['caching.enabled']:
-            # TODO: possible runtime shortcut here
-            return self.view(context, request)
+            response = self.view(context, request)
+            response.headers['X-View-Cache'] = 'DISABLED'
+            return response
 
         cache_manager = request.cache_manager
 
@@ -126,7 +192,20 @@ class ViewCacheDecorator(object):
 
         prefixes = [self.view.__module__, self.view.__name__]
 
-        return cache_manager.get_or_cache(get_result,
-                                          prefixes,
-                                          dependencies,
-                                          )
+        try:
+            result = cache_manager.get_or_cache(get_result,
+                                                prefixes,
+                                                dependencies)
+        except CacheDisabled:
+            return get_result()
+
+        response = result.data
+        result_info = result.info()
+        if result_info.hit:
+            request.registry.notify(ViewCacheHit(result_info.key, request))
+            response.headers['X-View-Cache'] = 'HIT'
+        else:
+            request.registry.notify(ViewCacheMiss(result_info.key, request))
+            response.headers['X-View-Cache'] = 'MISS'
+        response.headers['ETag'] = result.md5_hash()
+        return response