+++ pyramid_caching/tests/functional/test_metrics.py
@@ -16,21 +16,22 @@ class MetricsFunctionalTests(TestCase):
         session.add(User(id=1, name='Bob'))
         session.commit()
 
-    @mock.patch('pyramid_caching.ext.metrics.get_current_metrics')
-    def test_cache_miss(self, m_metrics):
+    @mock.patch('pyramid_metrics.utility.StatsClient')
+    def test_cache_miss(self, m_statsd):
         self.app.get('/users/1')
-        m_incr = m_metrics.return_value.incr
+        m_incr = m_statsd.return_value.incr
         m_incr.assert_called_once_with(
-            ('cache.miss', 'example.views:get_user'))
+            'cache.miss.example.views:get_user', count=1)
 
-    @mock.patch('pyramid_caching.ext.metrics.get_current_metrics')
-    def test_cache_hit(self, m_metrics):
+    @mock.patch('pyramid_metrics.utility.StatsClient')
+    def test_cache_hit(self, m_statsd):
         self.app.get('/users/1')
         self.app.get('/users/1')
-        m_incr = m_metrics.return_value.incr
+
+        m_incr = m_statsd.return_value.incr
         m_incr.assert_has_calls([
-            mock.call(('cache.miss', 'example.views:get_user')),
-            mock.call(('cache.hit', 'example.views:get_user')),
+            mock.call('cache.miss.example.views:get_user', count=1),
+            mock.call('cache.hit.example.views:get_user', count=1),
             ])